Revert "Add: manage 관련 layout과 페이지들 추가 #3"
