+++ app/[blog_name]/(category)/[category_name]/page.jsx
@@ -1,0 +1,7 @@
+/**
+ * 카테고리별 게시글 목록
+ * @params {category_name} 카테고리 이름
+ */
+export default function categoryArticles({ params }) {
+  return <div>카테고리별 게시글</div>;
+}
