+++ app/[blog_name]/manage/posts/page.jsx
@@ -1,13 +1,6 @@
 import React from "react";
-import ArticleList from "@components/ArticleList";
 
 /** 블로그 내 게시글관리페이지 */
-export default function posts({ params }) {
-  return (
-    <main className="main_by">
-      <div>
-        <ArticleList blog_name={params.blog_name} />
-      </div>
-    </main>
-  )
+export default function posts() {
+  return <div>글관리페이지입니다</div>;
 }
