+++ app/page.jsx
@@ -2,7 +2,7 @@ import React from "react";
 
 /* 메인페이지 - 사이트 소개페이지 */
 const Home = () => {
-  return <></>;
+  return <>ㅏㅏㅏ</>;
 };
 
 export default Home;
