+++ app/topic/culture/page.jsx
@@ -1,0 +1,3 @@
+export default function culture() {
+  return <div>culture</div>;
+}
