+++ app/topic/it/page.jsx
@@ -1,0 +1,3 @@
+export default function it() {
+  return <div>it</div>;
+}
