+++ app/topic/life/page.jsx
@@ -1,0 +1,3 @@
+export default function life() {
+  return <div>life</div>;
+}
