+++ app/topic/sports/page.jsx
@@ -1,0 +1,3 @@
+export default function sports() {
+  return <div>sports</div>;
+}
