+++ app/topic/travel/page.jsx
@@ -1,0 +1,3 @@
+export default function travel() {
+  return <div>travel</div>;
+}
